Basic LocalData Implementation
Take a look at providers/LocalData.ts
I'm being really careful here, and a lot of it is just empty method
stubs.
The idea is that LocalData would be the intermediary between WebAPI and
the Components. We could store globally accessible objects, and provide
getters and setters. It'll store things in LocalStorage, and make calls
to the API